breast_cancer_ai/requirements.txt
# ===============================================
# Breast Cancer AI Diagnostic System
# Requirements File
# ===============================================

# Core dependencies
streamlit>=1.28.0
biopython>=1.80
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchcam>=0.4.0

# Image Processing
opencv-python>=4.8.0

# NLP / NER (BioBERT entity extraction)
transformers>=4.30.0

# Knowledge Graph Visualization
networkx>=3.0
pyvis>=0.3.0

# Additional Visualization
wordcloud>=1.9.0

# Database
# sqlite3 is included in Python stdlib

# HTTP Requests (for Clinical Trials API)
requests>=2.28.0

# PDF Report Generation (optional)
reportlab>=4.0.0